Partial rewrite for better stability and better error reporting, by using a real parse & generate phase.

+++ terrarium_node.js
@@ -1,4 +1,4 @@
-var instrument = require('./instrument_node');
+var instrument = require('./instrument');
 var EventEmitter = require('events').EventEmitter;
 var util = require('util');
 var fs = require('fs');
@@ -13,45 +13,51 @@ function Terrarium() {
 util.inherits(Terrarium, EventEmitter);
 
 Terrarium.prototype.run = function(source) {
-  var instrumentedSource = instrument(source);
-  fs.writeFileSync(TMP, instrumentedSource.result);
-  this.child = child_process.fork(TMP, { silent: true });
+  try {
+    var instrumentedSource = instrument(source, 0, 'node');
+    fs.writeFileSync(TMP, instrumentedSource.source);
+    this.child = child_process.fork(TMP, { silent: true });
+    this.DATA = {};
+    var DATA = this.DATA;
+    var start = Date.now();
+    var TODO = instrumentedSource.TODO;
 
-  this.DATA = {};
-  var DATA = this.DATA;
-  var start = Date.now();
-  var TODO = instrumentedSource.TODO;
+    this.child.on('message', function(d) {
+      if (d && d.type === 'instrument') {
+        var name = d.name,
+          number = d.lineNumber,
+          val = d.value;
 
-  this.child.on('message', function(d) {
-    if (d && d.type === 'instrument') {
-      var name = d.name,
-        number = d.lineNumber,
-        val = d.value;
-
-      if (DATA[name + ':' + number] === undefined) {
-        DATA[name + ':' + number] = [];
+        if (DATA[name + ':' + number] === undefined) {
+          DATA[name + ':' + number] = [];
+        }
+        DATA[name + ':' + number].unshift({
+          name: name,
+          line: number,
+          val: val,
+          when: Date.now() - start
+        });
+        TODO[name + ':' + number] = true;
+        for (var k in TODO) {
+          if (!TODO[k]) return;
+        }
+        this.update();
       }
-      DATA[name + ':' + number].unshift({
-        name: name,
-        line: number,
-        val: val,
-        when: Date.now() - start
-      });
-      TODO[name + ':' + number] = true;
-      for (var k in TODO) {
-        if (!TODO[k]) return;
-      }
-      this.update();
-    }
-  }.bind(this));
+    }.bind(this));
 
-  this.child.stderr.on('data', function(d) {
-    this.emit('err', { error: d });
-  }.bind(this));
+    this.child.stderr.on('data', function(d) {
+      this.emit('err', { error: d });
+    }.bind(this));
 
-  this.child.on('exit', function(d) {
+    this.child.on('exit', function(d) {
+      this.emit('end');
+    }.bind(this));
+    return this;
+  } catch(e) {
+    this.emit('err', e);
     this.emit('end');
-  }.bind(this));
+    return this;
+  }
 };
 
 Terrarium.prototype.update = function() {
@@ -59,7 +65,7 @@ Terrarium.prototype.update = function() {
 };
 
 Terrarium.prototype.destroy = function() {
-  this.child.kill();
+  if (this.child) this.child.kill();
   if (fs.existsSync(TMP)) fs.unlinkSync(TMP);
 };
